diagnostics: Fix flaky CallGrpcMethodFromBackgroundThread
Fix flaky "DpslRequesterImplServerTest.CallGrpcMethodFromBackgroundThread"
tests that hung in case, due to some random timing inversion that caused
a run loop to never quit.

The issue was that the job that calls QuitEventLoop() (a job scheduled
from PerformRequest()) could be executed by the run loop that is spun by
DoSync(), which happens before RunEventLoop() is called. In such
scenario QuitEventLoop() is a no-op: the DPSL library, unlike libchrome,
chooses to make the event-loop-related helpers be more low-level.

The fix is to guard the RunEventLoop() call with a check of whether the
job was already executed.

BUG=b:233790719
TEST=run "cros_workon_make --board=$B diagnsotics-dpsl-test" 10000 times with stress-ng in background

Change-Id: If5f5444caf869a2d1e3a1ec309fdde9b8573892e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3738146
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Oleh Lamzin <lamzin@google.com>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>