camera: refactor CameraMonitor
It seems that the CameraMonitor thread's Attach(), Detach() and
StartMonitor() can be racy when there's Mojo connection error.

Refactor the CameraMonitor to start/stop the thread in ctor/dtor to
avoid the race. All the monitor tasks are run on the same thread to
avoid race conditions. While we're at it, consolidate the different
monitors and run all the monitor tasks on one thread.

BUG=b:237959427
TEST=`emerge-brya cros-camera` and manually tested on device.

Change-Id: I521effb5227226f9daa5950f53d4906e1edfe4ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3740370
Tested-by: Ricky Liang <jcliang@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>