screen-capture-utils: Improve kmsvnc responsivity to clicks
kmsvnc's main loop looks like this:

  while true:
    capture the screen and send changed pixels to clients
    wait for client events up to 16ms

Note that 16ms is a hard-coded value. It was picked so that, in an
ideally fast machine where capturing the screen finishes instantly,
the frame rate is 60fps. However in reality capturing the screen is
a costly operation; on kohaku it takes ~50ms, which yields ~15fps
(= 1000 / (50 + 16)).

The problem is that VNC server library processes only up to one
client event per loop. In other words, the server's event processing
rate is no more than the frame rate (~15/s on kohaku). When a client
sends events more frequently, events are queued in the server's buffer
and their processing are delayed.

For example, NoVNC is configured to send up to 60 events per second:
https://github.com/novnc/noVNC/blob/HEAD/core/rfb.js#L45
Hence, when the user moves the cursor, pointer events are piled up in
the server buffer and cause visible delays to the user interaction.

This patch resolves the problem by configuring the VNC server to
process events eagerly.

BUG=b:238138469
TEST=emerge-hatch screen-capture-utils
TEST=cros deploy --root=/usr/local DUT screen-capture-utils
TEST=http://go/ggzqi  # video recording

Change-Id: Ic7308b7d7841efb59c6c6954838654f3f9148cfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3746450
Tested-by: Shuhei Takahashi <nya@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>