shill: Remove --minimum-mtu flag
This flag is used for restricting the MTU value received from DHCP: if
it is smaller than the value set by this flag, the value from DHCP will
be ignored.

This flag was introduced in https://crrev.com/c/263522 for b/35547458.
Since we no longer need to support the Jetstream case and this flag is
not used elsewhere. We should be able to remove this flag safely.

This patch has the same effect with setting minimum_mtu to 576.

BUG=b:232177767
TEST=unit_tests

Change-Id: Ie8d73edccaf6f8099effc6a332c56eb7846b759d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3747538
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>