cryptohome: Leave backfilled KeyData empty
When loading a legacy Vault Keyset that has no KeyData stored in it,
backfill it with an empty KeyData.

This partially reverts crrev.com/c/3702820, in order to fix a
regression: inability to mount, when AuthSessions API is used, a kiosk
homedir that was created on Chrome OS 90 or earlier.

The underlying root cause is that the Chrome Login Screen implementation
for AuthSessions gets confused when it receives a KEY_TYPE_PASSWORD key
for a kiosk user, and eventually fails with the "Could not find Kiosk
key" error. This is indeed unexpected: kiosk keysets are special and
differ from regular password ones. We could fix it as either on Chrome
side (by force-overriding keyset types for all kiosk users) or on
Cryptohome side, and this CL proposes the latter. Basically, we stop
"guessing" the key type in Cryptohome, and let Chrome decide on this,
using its knowledge of user types and other information.

BUG=b:238126443
TEST=cros_workon_make --board=$B cryptohome
TEST=build CrOS 90 with tpm-simulator backported and ecryptfs enforced, enroll it (using hacks from crbug.com/895822), set up Kiosk user, snapshot user data and tpm-simulator's state, copy it to a ToT build with this CL cherry-picked, add "--enable-features=UseAuthsessionAuthentication" to /etc/chrome_dev.conf, verify the kiosk homedir can still be mounted

Change-Id: Ia8f0b54d24f76a65d5f544fe3f93400334730196
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3748199
Tested-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>