vm_tools: concierge: fix GPU-related option formatting
The --gpu and --gpu-render-server options were misusing the args array,
putting the whole option into the first part of the pair instead of
putting the option name (e.g. '--gpu') in the first part of the pair and
its parameters in the second part.

Fixes use of the deprecated --arg=options format.

BUG=b:235882579
TEST=Boot ARCVM and Crostini on hatch

Change-Id: I1c8e28d982e7133ad71fc0800e2a3b18fe723ebf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3749597
Reviewed-by: Chia-I Wu <olv@google.com>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>