hps: udev: start hpsd after relabelling /dev/cros-hps
In 50-hps.rules, the first rule matches the I2C peripheral and the
second rule matches the cros-hps device that gets bound to it.

The kernel will probe the I2C peripheral first and the cros-hps device
will be bound shortly after that, so the first udev rule will be
processed before /dev/cros-hps is ready.

But hpsd needs /dev/cros-hps to exist and be correctly labelled.
Move the command to start hpsd into the second rule, so that
/dev/cros-hps is ready to go when hpsd starts.

BUG=b:221356291
TEST=cros deploy $DUT hpsd  # then reboot DUT

Change-Id: I6e014d604bb0a09f1c842e0719df9d4711395ea5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3750269
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>