hps: make a failure to open /dev/cros-hps into a FATAL error
If there's something wrong with /dev/cros-hps (for example, wrong
SELinux context) we want hpsd to crash with a good message so we know
about it.

b/221356291 went largely unnoticed because the race condition happened
to be small enough that /dev/cros-hps became usable eventually. But
really hpsd expects /dev/cros-hps to always be usable and if it's ever
not, we should crash.

BUG=b:221356291
TEST=cros deploy $DUT hps  # reboot, also test with broken udev rules

Change-Id: I6fea26c9007f0d4f9fdf4731a982b4f8dd61075c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3750270
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>