common-mk: rebase source file path for install_config.
When a BUILD.gn in a subdirectory is referenced from another BUILD.gn,
the build root directory is not equal to the location where the BUILD.gn
is. Also, the working dir could be different during ebuild building
stage and installing stage.
This rebase the path to the system-absolute native style paths so it can
be used for any files and don't depend on the working directory.

BUG=b:231074077
TEST=checked with crrev.com/c/3631826. Make sure new file is installed.
TEST=To verify it works for non-workon packages, also test by changing
the CROS_WORKON_COMMIT in ebuild directly.

Change-Id: I950609c65a5132d85b133b8c8ac0b871646f1493
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3750674
Tested-by: Chung-Sheng Wu <chungsheng@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Chung-Sheng Wu <chungsheng@google.com>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>