ippusb_bridge: Reattach non-printer USB drivers
Currently ippusb_bridge detaches all kernel drivers when it gets ready
to claim the IPP-USB interfaces.  If a device exposes additional
non-printer interfaces, such as a USB mass storage passthrough, this
makes these interfaces unavailable to the host system.

At startup, we unconditionally set the device's active configuration to
whichever one includes the IPP-USB interfaces.  We do need to detach all
the drivers to switch the active USB configuration, so we still continue
to do that.  However, we can give the devices back to the rest of the
system by re-attaching the drivers for the interfaces we aren't
interested in.  Since this happens immediately after plugging in the
device, it shouldn't be any more disruptive to other processes on the
system than the current behavior.

Other processes on the system may begin using these returned interfaces,
so we want to avoid re-detaching them on subsequent interface claiming
loops unless the active configuration needs to be changed.  This should
avoid having a USB stick suddenly unmounted when starting a print or
scan job, for example.

The resulting behavior should be this:

At startup:
  1.  Loop over all the interfaces and detach the kernel drivers (same
      as the current behavior).
  2.  Set the active configuration to the one we want.
  3.  Re-attach the kernel drivers for non-printer interfaces (new).
  4.  Claim all the IPP-USB interfaces.
  5.  Release the IPP-USB interfaces after the idle timeout expires.

When a new connection comes:
  1.  If the active configuration is the one we want, skip detaching
      drivers.  Otherwise, do the same loop as the startup sequence.
  2.  Reclaim all the IPP-USB interfaces and pass traffic from the
      client.
  3.  Release the IPP-USB interfaces after the idle timeout expires.

As long as the MFP doesn't require switching configurations after
startup, this should result in sharing the non-printing interfaces with
the rest of the system.

BUG=b:238353195
TEST=Manually printed and scanned from test printers

Change-Id: Ia7512a2d6c88e3d42866ed16cd01fb57e91b65a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3752479
Reviewed-by: Mason Wilde <masonwilde@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>