fusebox: suppress kCloseMethod "not implemented" daemon error log
The server kCloseMethod is not implemented, but is currently called
in daemon file system Release code when MTP file reading ends.

The server response is always "not implemented" and the daemon adds
a LOG(ERROR) line to /var/log/messages as a result.

Add TODO: implement kCloseMethod and suppress the ERROR log for now
to avoid filling the log with known "not implemented" errors.

BUG=chromium:1249754,chromium:1327586
TEST=CQ and a manual test: read an MTP file and verify the server's
  kCloseMethod method is not called in file system Release code.

Change-Id: Ifa3e633b3d0e6c409227fea185d3ac035327775c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3753597
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Commit-Queue: Noel Gordon <noel@chromium.org>
Tested-by: Noel Gordon <noel@chromium.org>