shill: portal: move all configuration checks to Service
This patch regroups in Service::IsPortalDetectionDisabled all conditions
checked in Device::StartPortalDetection() pertaining to the
configuration of the currently selected Service or global settings in
the Manager. This allows to clarify expectations in Device unit tests
and will allow to properly know if portal detection should be stopped or
started when one of this configuration values change for the Service.

BUG=b:230339227
TEST=FEATURES=test emerge-$BOARD shill.

Change-Id: Iceb7d7b5d4287bfa9706e4178bd632d708cdeb6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3758983
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>