cryptohome: Pull CreateKeyBlobs up into the USS AddAuthFactor
This change takes the AuthFactor::CreateNew call that the
AddAuthFactorViaUserSecretStash does and inlines it. The ultimate goal
is to get the USS and Vault onto the same CreateKeyBlobs path (i.e. both
using the async version) but to do this we need the two paths to more
closely mirror each other.

The AuthFactor::CreateNew isn't used by anything else and so this change
also removes it. Fortunately, the abstraction it was providing was
minimal and so we're not losing much by getting rid of it.

BUG=b:236885089
TEST=FEATURES=test emerge-hatch cryptohome

Change-Id: Ic260709f5c7a312babf3b49520d84bb3e66bbbfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3758832
Auto-Submit: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>