vm_tools: concierge: supply options for battery and video args
This will let us remove the hacks in crosvm that allow deprecated
arguments that could previously either have options or not.

For example, previously either `--battery` or `--battery type=...` would
be allowed, but this creates an ambiguity when positional arguments such
as the kernel filename are allowed as well (we can't know whether an
unclaimed argument is meant to be the parameters for one of these
switch-or-option arguments or a positional argument).

The newly updated values correspond to the previous defaults when the
options were not specified.

BUG=b:235882579
TEST=Start Crostini and ARCVM on hatch

Change-Id: I9ee6bbe0c2cc1ea4b8188400c3c273afc74609e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3759267
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>