authpolicy: Update usage of generate_policy_source.py
In crrev/c/3714518 a refactoring of the generate_policy_source.py script
was performed. This script now expects an auto-generated real JSON file
instead of the pseudo-json go/policy_templates.

This CL updates the BUILD.gn files to use the new auto-generated file as
the input for generate_policy_source.py.

The generation of the new file is done in the protofiles package, that
authpolicy depends on.

BUG=chromium:1294298
TEST=FEATURES=test emerge-eve authpolicy

Cq-Depend: chromium:3762549
Change-Id: I3c7d76e7a19bf329a91e55c829a5e53e50bcf650
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3762581
Commit-Queue: Jana Grill <janagrill@google.com>
Reviewed-by: Felipe Andrade <fsandrade@chromium.org>
Reviewed-by: Ram Chandrasekar <rchandrasekar@google.com>
Tested-by: Jana Grill <janagrill@google.com>