cryptohome: add oom score lines to all cryptohome upstart scripts
This adds an "oom score" to all of the (6) upstart scripts in cryptohome
that were missing such a declaration. This silences a warning-error in
the build, but more generally addresses the constraint specified in
https://www.chromium.org/chromium-os/chromiumos-design-docs/boot-design/#runtime-resource-limits
that every service needs an oom score.

I used a default value of -100 because this is specified as "CrOS
daemons that can recover" which seemed appropriate. However, for
cryptohomed-client I used -200 since losing it would impact unmounting
drives, and cryptohomed is -500 to put it on par with ui.conf.

Note that some of these conf scripts report other pre-existing lint
errors which this change does not attempt to address, as they're
unrelated to the oom score (or lack thereof).

BUG=b:239039956
TEST=FEATURES=test emerge-hatch cryptohome
TEST=booted image in vm, tested login & logout

Change-Id: Ibcfea3342dd96a3dcd554ae642facf8b143d05ba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3764020
Reviewed-by: Allen Webb <allenwebb@google.com>
Auto-Submit: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>