shill: pseudonymize session tags
To add an extra layer to avoid leaking session tags to logs, we
pseudonymize those tags when they're logged by adding a salt and hashing
the result.
On startup shill generates a 32-byte salt that is never logged. When
the session tag is logged, we first concatenate that random salt with
the session tag, then SHA256() the resulting string. The first 8 bytes
of the resulting hash are then used to write to the log.

Even with that extra layer of pseudonymization, we still keep the log
verbosity higher than what could be set in verified mode.

BUG=b:226145383
TEST=cros-workon-${BOARD} start shill
TEST=FEATURES=test emerge-${BOARD} shill
TEST=emerge-${BOARD} shill
TEST=cros deploy ${DUT} shill
TEST=restart shill SHILL_LOG_SCOPES=wifi+device+metrics
SHILL_LOG_LEVEL=-4

Change-Id: Ibf4c4d06e0d61d2e5cfc3d123176f8cbf380ad49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3763399
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>