power: Change Adaptive Charging metric states
This makes it such that the state for Adaptive Charging can be
HEURISTIC_DISABLED, even when the feature is not enabled. This means
that when the state is USER_DISABLED or NOT_SUPPORTED, the heuristic
check passes. This allows for more useful metrics for reporting how well
the ML model does.

BUG=chromium:1243163
TEST=unittests

Change-Id: Ifa1a3d60eb3929cbee76c6460bef44b6a1ac5447
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3764889
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>