hwsec-test-utils: Ensure TPM owned in cross_version
Add an extra initialization step that makes sure the TPM becomes owned
in the prepare_cross_version_login_data.py test.

This is only needed when running the script for old, pre-M96 images,
because on these images the ToT Tast framework fails to establish the
ownership itself (as it depends on the functionality added in
b/193200420). If we don't make the TPM owned, the snapshots created by
the test won't have TPM-backed keysets, which introduced a subtle
testing gap.

BUG=b:237120336,b:238006044
TEST=run "prepare_cross_version_login_data.py --board=betty --version=R91-13904.0.0 --output-dir=.",
     check "cryptohome --action=tpm_status" says the TPM becomes owned while
     this runs, parse the resulting keyset files and check they don't have
     "flags: 2" keysets

Change-Id: I9a4223683b18516d7f9d536c165e55ec28e62854
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3755841
Reviewed-by: Leo Lai <cylai@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Ching-Kang Yen <chingkang@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>