shill: Static IPv6 path in Cellular uses Network::Start()
This patch change the static IPv6 code path to use Network::Start().
For the code removed in Cellular:
- SelectService() and SetServiceState() will always be called in the
  following part of this function.
- StartIPv6() and ConfigureStaticIPv6Address() will be called in
  Network::Start().

Also change the logic in Network::Start(), since it can be called with
a IPv6-only setup. One visible result of this change is that, an
immediate DHCP failure will stop the network directly before, which is
no longer the case now.

BUG=b:232177767
TEST=test_that -b octopus-kernelnext $DUT cellular_Smoke.verizon
TEST=manually checked the dual-stack connectivity (by ping) on trogdor
  and hatch with T-Mobile in the lab.
TEST=FEATURES="test" emerge-${BOARD} shill

Change-Id: I9ce7e5f3c4c0a3b62263803126efbb4742de43fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3777083
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>