cryptohome: Refactor Authenticate via USS in preparation for async
This applies similar refactors to AuthenticateViaUserSecretStash that
were recently made to AddAuthFactorViaUserSecret stash, to move it
towards being able to use the async DeriveKeyBlobs call.

The changes being:
 - pull the derive call up into auth session
 - push all of the post-derive code down into a separate function
 - switch the functions from returns status to ReplyWithError+on_done

In this case I was able to make use of LoadUSSMainKeyAndFsKeyset as the
intended DeriveCallback implementation. This did also require
rearranging some of its parameters in preparation for it becoming that.

BUG=b:236885089
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I4010815fbd659994b075f6378375b4ba3a626750
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3774514
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Auto-Submit: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>