ml_core: Adding no_effect and set_effect callback
There was no mechanism by which we could reset effects back to no
effect. The mechanism by which we received feedback on whether
SetEffect was successful was not sufficient.

Can now use kNone EffectType to indicate that all effects should be
reset, and can supply a callback to the SetEffect method which will
receive the result of the SetEffect call.

BUG=b:239747989
TEST=Ran on Volteer and confirmed behaviour with g3 shared library

Change-Id: Ic7c9b85199c292218cacf12f0987128b1b8e878a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3775115
Reviewed-by: Tom Shafron <shafron@google.com>
Reviewed-by: Michael Olney <mpolney@google.com>
Commit-Queue: Tom Shafron <shafron@google.com>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Tested-by: Tom Shafron <shafron@google.com>