permission_broker: Change internal allow list to fixed allow list
Currently the permission broker assumes fixed devices without a
removable parent are internal and will deny them if they are not in the
internal allow list. To account for external devices marked as fixed
which don't register a removable parent, this CL changes the internal
allow list to a fixed allow list. It also adds the Citrix Nuance
PowerMic III to the allow list which is an external device marked as
fixed which does not register a removable parent on older kernel
versions.

BUG=b:239707364
TEST=cros_workon_make --board=brya chromeos-base/permission_broker
--test

Change-Id: Ia7f6a34a18904afb9973b81ea4e47ca2c1b83bfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3777356
Commit-Queue: Jameson Thies <jthies@google.com>
Tested-by: Jameson Thies <jthies@google.com>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>