parallax: Create Streaming client
Adds a client to stream samples from the server. This client
can connect to a sample server and fetch data periodically.
If a server does not respond to the requests, it will continue
to attempt until a timeout expires which will disconnect the
client.

Adds logic to only fetch subsets of samples to reduce amount
of data transferred.

BUG=b:239704535
TEST=`python frontend_test.py` and `npm run build -- report`
    Validated connection and disconnections are handled.
    Validated streaming samples.

Change-Id: I9188469f46c08657879cab819e66cfe8c56cf1d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3778357
Commit-Queue: Brian Nemec <bnemec@google.com>
Tested-by: Brian Nemec <bnemec@google.com>
Reviewed-by: Ziting Shen <zitingshen@google.com>