vm_tools: Use regex to match concerge's wayland socket
In a future CL, we will be enforcing that the wayland socket exists and
is not a symlink, which involves changing the way the tests work. Part
of that will be allowing the wayland regex to be overridden with a
custom regex.

We do the regex in advance, in this CL, mostly to show that the current
set of tests passes unmodified despite moving to regex.

BUG=b:237960042
TEST=FEATURES=test emerge-volteer vm_host_tools

Change-Id: I55f5cb6c45d0a5557a359bcd85d7627c645baed6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3777086
Auto-Submit: Nic Hollingum <hollingum@google.com>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Tested-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>