libec: Implement EC_CMD_FORCE_LID_OPEN in libec
power_manager is using ectool to set forcelidopen. ectool can hang on
reven devices when kernel lockdown is off so it is better to use
libec instead. This CL just adds the command and unit tests to libec.

I tested the functionality by deploying a power_manager edited to set
forcelidopen to 1 via the libec command during initialization. I
confirmed that the screen stayed on even as the lid was closed just as
it would from the ectool command. I then loaded a power_manager that set
force_lid_open to 0 and verified that the screen turned off when the lid
was closed after restarting powerd.

BUG=b:199548301, b:203042528
TEST=cros_run_unit_tests --board zork --packages libec && cros deploy
${DUT} libec power_manager && restart powerd && manually close lid /
compare with behavior of ectool forcelidopen 1

Change-Id: I663bced506f6e0b9c97c583433f0b23bdecbd15d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3779704
Tested-by: Patrick Thompson <ptf@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Patrick Thompson <ptf@google.com>