hps: delete 'hps boot' sub-command
It's only possible to use the hps command as a debugging tool with the
kernel driver unbound. But when the kernel driver is unbound, the hps
command can't power cycle the peripheral so the boot cycle doesn't work.

The best way to test the boot cycle is to either let hpsd run through it
normally, or do all the individual steps by hand using the hps command.

BUG=b:228282193
TEST=FEATURES=test emerge-brya hps-tool
TEST=/mnt/host/source/src/platform2/common-mk/platform2_test.py \
     --sysroot=/build/brya -- hps --help

Change-Id: Ia99f71e31706b9adf0e509f64453656fba37a02b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3782464
Commit-Queue: David Lattimore <dml@chromium.org>
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Auto-Submit: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: David Lattimore <dml@chromium.org>