sirenia: loosen MMS checks on balloon deltas
When balloon deltas would deflate below zero/inflate above mem_size,
clamp the deltas instead of rejecting the request. If the total delta
across all balloons is not zero, adjust the list of deltas so that it
becomes zero instead of rejecting the request. This is done in
preparation for allowing MMS to make its own adjustments to balloons,
since such adjustments could conflict with requests from the guest.

BUG=b:229176768
TEST=unit tests; boot ManaTEE and manually test w/sibling VMs

Change-Id: Id0aa782ec46018a20769f95ed0f981fa62b3fde6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3807194
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: David Stevens <stevensd@chromium.org>
Tested-by: David Stevens <stevensd@chromium.org>