cryptohome: Update bind mount migration logic to be atomic
The existing logic for bind mount migration moved each top level item in
the ~/Downloads folder to ~/MyFIles/Downloads. This instead implements
the logic outlined in the DD [1] to move the entire folder. This also
adds in additional UMAs to track migration status and a separate test to
validate the inverse migration in the case of a rollback.

[1] go/remove-downloads-bind-mount-dd

BUG=b:239899266, b:239899057
TEST=cros_run_unit_tests --board hatch --packages cryptohome
TEST=Deploy cryptohome and verify UMA and bind mount migration

Change-Id: I6a21e77f010d32414108c8082e1fca392989776e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3782466
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>
Tested-by: Ben Reich <benreich@chromium.org>