runtime_probe: Remove factory_runtime_probe_installer
We can call factory runtime probe directly in factory flow. Thus we
don't need to bundle it into factory toolkit. It is also seldom for
partners and factory owner to manually create a static linked binary. If
we want to have temporary patch we can have a CL and build a image based
on it.

BUG=b:236666852
TEST=CQ

Change-Id: I303a2ba168dbf83ca59be7ac607b85c0b9dc2d19
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3779821
Reviewed-by: Yong Hong <yhong@chromium.org>
Tested-by: Chung-Sheng Wu <chungsheng@google.com>
Commit-Queue: Chung-Sheng Wu <chungsheng@google.com>