login: Reset only block_devmode when FrE is not required
The FrE not required means the devmode is not blocked. So, it makes
sense to update the block_devmode to zero. The check_enrollment flag
however should remain unchanged. This is needed for the case the
device is new and needs to go through ZTE. If ZTE fails, it should
still be able to retry later. This will allow for any mistakes in
ZTE data to be fixed and make the device retry successfully later.

BUG=b:237003956
TEST=Manual test on device and unit test

Change-Id: I9edc1a972f9b77688b3be946096915c6a28ecfc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3805301
Commit-Queue: Igor <igorcov@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Igor <igorcov@chromium.org>