dns-proxy: Always clear watchers before ares processing
Watching a closed file descriptors is discouraged as it can lead
to a dangerous race-condition i.e. if a new fd is created with
the same fd number. For the same reason, FileDescriptorWatcher
can cause a CHECK crash when dealing with a closed file descriptor.

Ares might close its fds on ares_process_fd. This patch makes it
so that the watchers are always cleared prior to calling any ares
processing.

BUG=b:239787406
TEST=tast run <DUT> network.DNSProxy.doh_off
TEST=dig +tcp <URL> any

Change-Id: Ic18670c8717443c1e379928bddfcf89cb78f2007
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3780506
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>