shill: cellular: Prioritize the LastGoodAPN when it's on MODB
When a user never chooses an APN in the UI, shill will use round robin
to find the correct APN every time. This means that shill will report
InvalidAPN multiple times when the APN is not the first in the list.
This also means that the registration with the carrier will take a very
long time when multiple attach APNs exist in the MODB.
Restructure the try list so that the Modem APNs are added
first(no change in logic), followed by the LastGoodAPN if and only if
the APN is also present in the MODB, then followed by the rest of the
APNs in the MODB.

BUG=b:239989259
TEST=Test multiple APN configurations with the callbox.

Change-Id: I1e615bf365e1388addf8e228a1eda30ca3c0f9dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3783577
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>