power: Fix failed hibernate reporting
The "error=" line after "hiberman hibernate" masks the return value,
making it look like hibernate succeeded always. Fix the order here so
hiberman's return code is what's actually evaluated, and add an
informative error print so it's visible in the logs.

Note that the "error writing to /sys/power/state" message is not
technically correct anymore, but it's used all over the place by tests
and such, so IMO it's better not to mess with it just for hibernate.

BUG=b:232184089
TEST=perform a failed hibernation, see powerd retry.

Change-Id: I8b82cc2ae9ad94820e431185e247143adc19ec43
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3783416
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>