power: Differentiate failed hibernate in metrics/logging
Pass down whether or not it's a failed hibernation (versus a failed
suspend) that causes the system to shut down. This shows up both in
syslog as the "reason" for shutdown, as well as in the histograms for
why we shut down. Hibernate and suspend failures are likely to be caused
by different bugs, so it makes sense to track them differently.

The log line looks something like this:
powerd: [daemon.cc(1720)] Shutting down, reason: suspend-failed

BUG=b:232184089
TEST=Hobble hibernate and attempt it, observe shutdown

Change-Id: I5e16e40c50dbc6f61b54d1f6d7a80b8363838dcc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3783417
Tested-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>