shill: cellular: Preserve the APN version.
Currently, the LastGoodAPN is stored in shill's default profile without
the version number,and the APN is loaded without loading the version
from the profile. When attach APNs exist, a ReAttach triggers a service
destruct/construct, and the values are not loaded.
This is almost harmless when no attach APNs are used, or when the user
chooses an APN in the UI, but when attach APNs exist or the user expects
an automatic connection, this has 2 effects:
- The LastGoodAPN cannot be used to prioritize a list of attach APNs,
  causing long registration times.
- If multiple APNs exist in the DB, shill will send an InvalidAPN
  metric for each of them until the correct APN is used.

BUG=b:239989259
TEST=Test with multiple APN configurations using the callbox.

Change-Id: I8aebc3d7fc764c73687c13a756ae30f2f320cf4d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3783576
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Pavan Holla <pholla@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>