cryptohome: eliminate explicit instantiations in error converters
This is a refactor that has limited functional impact but which removes
what is generally considered to be an anti-pattern in C++, defining a
(public) template function that's implemented in a .cc file with
explicit instantiations.

In this case I applied the most standard solution to header templates,
which is to make them inline functions (i.e. putting the definitions in
the headers). While inline functions are not super popular in the
Chromium style they're still fairly common, and neither of these
functions are hugely complex so we're not pulling a large amount of code
into the header.

I also made a slight adjustment to the ReplyWithError code to make the
reply parameter a value parameter instead of a const reference (while
eliminating the internal copy). This produces the same end result but is
more optimal. The PopulateRelyWithError is also slightly tweaked to more
efficiently move the protobuf objects.

BUG=b:240152277
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I69a2bb9661a5e671cd7fb2a3dfd0e98759d3d966
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3784761
Reviewed-by: Yi Chou <yich@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: John L Chen <zuan@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Auto-Submit: John Admanski <jadmanski@chromium.org>