hps: tests: inject transient boot faults as register read failures
These tests are covering the "transient" boot fault scenario, where
a register read or write would fail at a seemingly-random point in the
boot flow, due to suspend-and-resume.

Currently the tests are simulating a failure of the reset command, which
we did see occasionally in the field. But a much more common failure was
reading the status register during CheckStage0() or CheckStage1(). Make
the tests simulate that failure mode instead.

This prepares for a follow-up commit which will stop sending the reset
command, thus these tests will no longer be able to inject a failure at
that point.

BUG=b:240226816
TEST=FEATURES=test emerge-brya hpsd hps-tool

Change-Id: Ia26e11620f864ee9e84880f1406c77cd3cd96a85
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3784522
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>